+++ sdk/request/request.go
@@ -6,15 +6,15 @@ import (
 	"log"
 	"net/http"
 	"net/url"
-	"regexp"
 	"time"
 
-	"github.com/terraform-providers/terraform-provider-fortiflexvm/sdk/config"
+	auth "github.com/terraform-providers/terraform-provider-fortiflexvm/sdk/auth"
 )
 
 // Request describes the request to FortiFlex service
 type Request struct {
-	Config       config.Config
+	Auth         *auth.Auth
+	HTTPCon      *http.Client
 	HTTPRequest  *http.Request
 	HTTPResponse *http.Response
 	Path         string
@@ -22,21 +22,22 @@ type Request struct {
 	Data         *bytes.Buffer
 }
 
-// New creates request object with http method, path, params and data,
+// NewRequest creates request object with http method, path, params and data,
 // It will save the http request, path, etc. for the next operations
 // such as sending data, getting response, etc.
 // It returns the created request object to the gobal plugin client.
-func New(c config.Config, method string, path string, params interface{}, data *bytes.Buffer) *Request {
+func NewRequest(author *auth.Auth, httpcon *http.Client, method string, path string, params interface{}, data *bytes.Buffer) *Request {
 	var h *http.Request
 
-	if data == nil {
+	if data == nil { // This "if-else" is necessary
 		h, _ = http.NewRequest(method, "", nil)
 	} else {
 		h, _ = http.NewRequest(method, "", data)
 	}
-
+	h.Header.Set("Content-Type", "application/json")
 	r := &Request{
-		Config:      c,
+		Auth:        author,
+		HTTPCon:     httpcon,
 		Path:        path,
 		HTTPRequest: h,
 		Params:      params,
@@ -47,22 +48,15 @@ func New(c config.Config, method string, path string, params interface{}, data *
 
 // Send request data to FortiFlex.
 // If errors are encountered, it returns the error.
-func (r *Request) Send() error {
-	return r.SendRequest(5)
-}
-
-// SendRequest request data to FortiFlex.
-// If errors are encountered, it returns the error.
-func (r *Request) SendRequest(retries int) error {
-	r.HTTPRequest.Header.Set("Content-Type", "application/json")
-	u := buildURL(r)
+func (r *Request) Send(retries int) error {
+	u := "https://support.fortinet.com" + r.Path
 
 	var err error
-	if r.Config.Auth.Token == "" {
+	if r.Auth.Token == "" {
 		err = fmt.Errorf("Could not find a API Token!")
 		return err
 	}
-	var bearer = "Bearer " + r.Config.Auth.Token
+	var bearer = "Bearer " + r.Auth.Token
 	r.HTTPRequest.Header.Set("Authorization", bearer)
 	r.HTTPRequest.URL, err = url.Parse(u)
 	if err != nil {
@@ -72,9 +66,9 @@ func (r *Request) SendRequest(retries int) error {
 
 	retry := 0
 	for retry <= retries {
-		rsp, err := r.Config.HTTPCon.Do(r.HTTPRequest)
+		rsp, err := r.HTTPCon.Do(r.HTTPRequest)
 		if err != nil {
-			log.Printf("Error found: %v, will resend again %s, %d", filterapikey(err.Error()), u, retry)
+			log.Printf("[ERROR] Request '%s' | %v | retry time %d", u, err.Error(), retry)
 		} else {
 			r.HTTPResponse = rsp
 			break
@@ -86,18 +80,4 @@ func (r *Request) SendRequest(retries int) error {
 		err = fmt.Errorf("Can't connect to server, please try it later.")
 	}
 	return err
-}
-
-func filterapikey(v string) string {
-	re, _ := regexp.Compile("access_token=.*?\"")
-	res := re.ReplaceAllString(v, "access_token=***************\"")
-
-	return res
-}
-
-func buildURL(r *Request) string {
-	u := "https://support.fortinet.com"
-	u += r.Path
-
-	return u
 }