+++ sdk/sdkcore/sdkfaz.go
@@ -1,9 +1,8 @@
-// Copyright 2020 Fortinet, Inc. All rights reserved.
-// Author: Xing Li (@lix-fortinet), Hongbin Lu (@fgtdev-hblu)
-// Documentation:
-// Xing Li (@lix-fortinet), Hongbin Lu (@fgtdev-hblu),
+// Copyright 2023-2024 Fortinet, Inc. All rights reserved.
+// Author: Xinwei Du (@dux-fortinet), Xing Li (@lix-fortinet), Hongbin Lu (@fgtdev-hblu)
+// Documentation: Xinwei Du (@dux-fortinet), Xing Li (@lix-fortinet), Hongbin Lu (@fgtdev-hblu)
 
-// Description: Description: SDK for FortiFlex Provider
+// Description: SDK for FortiFlex Provider
 
 package forticlient
 
@@ -11,22 +10,13 @@ import (
 	"fmt"
 )
 
-// type creatUpdateOutput struct {
-// 	Vdom       string  `json:"vdom"`
-// 	Mkey       string  `json:"mkey"`
-// 	Status     string  `json:"status"`
-// 	HTTPStatus float64 `json:"http_status"`
-// }
-
 // ReadProgramsList API operation for FortiFlex gets the Programs list
 // Returns the requested value when the request executes successfully.
 // Returns error for service API and SDK errors.
 func (c *FortiSDKClient) ReadProgramsList(params *map[string]interface{}) (mapTmp map[string]interface{}, err error) {
-	HTTPMethod := "POST"
 	path := "/ES/api/fortiflex/v2/programs/list"
 	rspKey := "programs"
-
-	mapTmp, err = read(c, HTTPMethod, path, true, rspKey, params)
+	mapTmp, err = read(c, "POST", path, rspKey, params)
 	return
 }
 
@@ -34,11 +24,9 @@ func (c *FortiSDKClient) ReadProgramsList(params *map[string]interface{}) (mapTm
 // Returns the requested value when the request executes successfully.
 // Returns error for service API and SDK errors.
 func (c *FortiSDKClient) ReadConfigsList(params *map[string]interface{}) (mapTmp map[string]interface{}, err error) {
-	HTTPMethod := "POST"
 	path := "/ES/api/fortiflex/v2/configs/list"
 	rspKey := "configs"
-
-	mapTmp, err = read(c, HTTPMethod, path, true, rspKey, params)
+	mapTmp, err = read(c, "POST", path, rspKey, params)
 	return
 }
 
@@ -46,11 +34,9 @@ func (c *FortiSDKClient) ReadConfigsList(params *map[string]interface{}) (mapTmp
 // Returns the requested value when the request executes successfully.
 // Returns error for service API and SDK errors.
 func (c *FortiSDKClient) CreateConfig(params *map[string]interface{}) (mapTmp map[string]interface{}, err error) {
-	HTTPMethod := "POST"
 	path := "/ES/api/fortiflex/v2/configs/create"
 	rspKey := "configs"
-
-	mapTmp, err = createUpdate(c, HTTPMethod, path, rspKey, params)
+	mapTmp, err = createUpdate(c, "POST", path, rspKey, params)
 	return
 }
 
@@ -58,11 +44,9 @@ func (c *FortiSDKClient) CreateConfig(params *map[string]interface{}) (mapTmp ma
 // Returns the requested value when the request executes successfully.
 // Returns error for service API and SDK errors.
 func (c *FortiSDKClient) UpdateConfig(params *map[string]interface{}) (mapTmp map[string]interface{}, err error) {
-	HTTPMethod := "POST"
 	path := "/ES/api/fortiflex/v2/configs/update"
 	rspKey := "configs"
-
-	mapTmp, err = createUpdate(c, HTTPMethod, path, rspKey, params)
+	mapTmp, err = createUpdate(c, "POST", path, rspKey, params)
 	return
 }
 
@@ -70,11 +54,9 @@ func (c *FortiSDKClient) UpdateConfig(params *map[string]interface{}) (mapTmp ma
 // Returns the requested value when the request executes successfully.
 // Returns error for service API and SDK errors.
 func (c *FortiSDKClient) UpdateConfigStatus(params *map[string]interface{}, op string) (mapTmp map[string]interface{}, err error) {
-	HTTPMethod := "POST"
 	path := fmt.Sprintf("/ES/api/fortiflex/v2/configs/%v", op)
 	rspKey := "configs"
-
-	mapTmp, err = createUpdate(c, HTTPMethod, path, rspKey, params)
+	mapTmp, err = createUpdate(c, "POST", path, rspKey, params)
 	return
 }
 
@@ -82,11 +64,9 @@ func (c *FortiSDKClient) UpdateConfigStatus(params *map[string]interface{}, op s
 // Returns the requested value when the request executes successfully.
 // Returns error for service API and SDK errors.
 func (c *FortiSDKClient) ReadEntitlementsList(params *map[string]interface{}) (mapTmp map[string]interface{}, err error) {
-	HTTPMethod := "POST"
 	path := "/ES/api/fortiflex/v2/entitlements/list"
 	rspKey := "entitlements"
-
-	mapTmp, err = read(c, HTTPMethod, path, true, rspKey, params)
+	mapTmp, err = read(c, "POST", path, rspKey, params)
 	return
 }
 
@@ -94,11 +74,9 @@ func (c *FortiSDKClient) ReadEntitlementsList(params *map[string]interface{}) (m
 // Returns the requested value when the request executes successfully.
 // Returns error for service API and SDK errors.
 func (c *FortiSDKClient) CreateEntitlementsVM(params *map[string]interface{}) (mapTmp map[string]interface{}, err error) {
-	HTTPMethod := "POST"
 	path := "/ES/api/fortiflex/v2/entitlements/vm/create"
 	rspKey := "entitlements"
-
-	mapTmp, err = createUpdate(c, HTTPMethod, path, rspKey, params)
+	mapTmp, err = createUpdate(c, "POST", path, rspKey, params)
 	return
 }
 
@@ -106,11 +84,9 @@ func (c *FortiSDKClient) CreateEntitlementsVM(params *map[string]interface{}) (m
 // Returns the requested value when the request executes successfully.
 // Returns error for service API and SDK errors.
 func (c *FortiSDKClient) CreateEntitlementsHW(params *map[string]interface{}) (mapTmp map[string]interface{}, err error) {
-	HTTPMethod := "POST"
 	path := "/ES/api/fortiflex/v2/entitlements/hardware/create"
 	rspKey := "entitlements"
-
-	mapTmp, err = createUpdate(c, HTTPMethod, path, rspKey, params)
+	mapTmp, err = createUpdate(c, "POST", path, rspKey, params)
 	return
 }
 
@@ -118,11 +94,9 @@ func (c *FortiSDKClient) CreateEntitlementsHW(params *map[string]interface{}) (m
 // Returns the requested value when the request executes successfully.
 // Returns error for service API and SDK errors.
 func (c *FortiSDKClient) CreateEntitlementsCloud(params *map[string]interface{}) (mapTmp map[string]interface{}, err error) {
-	HTTPMethod := "POST"
 	path := "/ES/api/fortiflex/v2/entitlements/cloud/create"
 	rspKey := "entitlements"
-
-	mapTmp, err = createUpdate(c, HTTPMethod, path, rspKey, params)
+	mapTmp, err = createUpdate(c, "POST", path, rspKey, params)
 	return
 }
 
@@ -130,11 +104,9 @@ func (c *FortiSDKClient) CreateEntitlementsCloud(params *map[string]interface{})
 // Returns the requested value when the request executes successfully.
 // Returns error for service API and SDK errors.
 func (c *FortiSDKClient) ReadEntitlementsPoint(params *map[string]interface{}) (mapTmp map[string]interface{}, err error) {
-	HTTPMethod := "POST"
 	path := "/ES/api/fortiflex/v2/entitlements/points"
 	rspKey := "entitlements"
-
-	mapTmp, err = read(c, HTTPMethod, path, true, rspKey, params)
+	mapTmp, err = read(c, "POST", path, rspKey, params)
 	return
 }
 
@@ -142,11 +114,9 @@ func (c *FortiSDKClient) ReadEntitlementsPoint(params *map[string]interface{}) (
 // Returns the requested value when the request executes successfully.
 // Returns error for service API and SDK errors.
 func (c *FortiSDKClient) UpdateVmUpdate(params *map[string]interface{}) (mapTmp map[string]interface{}, err error) {
-	HTTPMethod := "POST"
 	path := "/ES/api/fortiflex/v2/entitlements/update"
 	rspKey := "entitlements"
-
-	mapTmp, err = createUpdate(c, HTTPMethod, path, rspKey, params)
+	mapTmp, err = createUpdate(c, "POST", path, rspKey, params)
 	return
 }
 
@@ -154,11 +124,9 @@ func (c *FortiSDKClient) UpdateVmUpdate(params *map[string]interface{}) (mapTmp 
 // Returns the requested value when the request executes successfully.
 // Returns error for service API and SDK errors.
 func (c *FortiSDKClient) UpdateVmUpdateStatus(params *map[string]interface{}, op string) (mapTmp map[string]interface{}, err error) {
-	HTTPMethod := "POST"
 	path := fmt.Sprintf("/ES/api/fortiflex/v2/entitlements/%v", op)
 	rspKey := "entitlements"
-
-	mapTmp, err = createUpdate(c, HTTPMethod, path, rspKey, params)
+	mapTmp, err = createUpdate(c, "POST", path, rspKey, params)
 	return
 }
 
@@ -166,11 +134,9 @@ func (c *FortiSDKClient) UpdateVmUpdateStatus(params *map[string]interface{}, op
 // Returns the requested value when the request executes successfully.
 // Returns error for service API and SDK errors.
 func (c *FortiSDKClient) UpdateVmUpdateRegenerateToken(params *map[string]interface{}) (mapTmp map[string]interface{}, err error) {
-	HTTPMethod := "POST"
 	path := "/ES/api/fortiflex/v2/entitlements/vm/token"
 	rspKey := "entitlements"
-
-	mapTmp, err = createUpdate(c, HTTPMethod, path, rspKey, params)
+	mapTmp, err = createUpdate(c, "POST", path, rspKey, params)
 	return
 }
 
@@ -178,7 +144,6 @@ func (c *FortiSDKClient) UpdateVmUpdateRegenerateToken(params *map[string]interf
 // Returns the requested value when the request executes successfully.
 // Returns error for service API and SDK errors.
 func (c *FortiSDKClient) ReadGroupsList(params *map[string]interface{}) (mapTmp map[string]interface{}, err error) {
-	HTTPMethod := "POST"
 	path := "/ES/api/flexvm/v1/groups/list"
 	rspKey := "groups"
 	if value, ok := (*params)["accountId"]; ok {
@@ -186,8 +151,7 @@ func (c *FortiSDKClient) ReadGroupsList(params *map[string]interface{}) (mapTmp 
 			path = "/ES/api/fortiflex/v2/groups/list"
 		}
 	}
-
-	mapTmp, err = read(c, HTTPMethod, path, true, rspKey, params)
+	mapTmp, err = read(c, "POST", path, rspKey, params)
 	return
 }
 
@@ -195,10 +159,8 @@ func (c *FortiSDKClient) ReadGroupsList(params *map[string]interface{}) (mapTmp 
 // Returns the requested value when the request executes successfully.
 // Returns error for service API and SDK errors.
 func (c *FortiSDKClient) ReadGroupsNexttoken(params *map[string]interface{}) (mapTmp map[string]interface{}, err error) {
-	HTTPMethod := "POST"
 	path := "/ES/api/fortiflex/v2/groups/nexttoken"
 	rspKey := "entitlements"
-
-	mapTmp, err = read(c, HTTPMethod, path, true, rspKey, params)
+	mapTmp, err = read(c, "POST", path, rspKey, params)
 	return
 }