+++ sdk/sdkcore/sdkutils.go
@@ -7,80 +7,64 @@ import (
 	"io/ioutil"
 	"log"
 	"time"
+
+	request "github.com/terraform-providers/terraform-provider-fortiflexvm/sdk/request"
 )
 
-func createUpdate(c *FortiSDKClient, method string, path string, rspKey string, params *map[string]interface{}) (mapTmp map[string]interface{}, err error) {
-	result, string_body, err := sendRequest(c, method, path, params)
+func createUpdate(client *FortiSDKClient, method string, path string, rspKey string, params *map[string]interface{}) (map[string]interface{}, error) {
+	result, string_body, err := sendRequest(client, method, path, params)
 	if err != nil {
 		return nil, err
 	}
-
 	err = fortiAPIErrorFormat(result, string_body)
-	if err == nil {
-		if rspKey == "entitlements" && result[rspKey] == nil { // TMP FortiFlex API BUG fix
-			rspKey = "vms"
-		}
-		if result[rspKey] != nil {
-			if confMap, ok := result[rspKey].(map[string]interface{}); ok {
-				mapTmp = confMap
-				return
-			} else if confList, ok := result[rspKey].([]interface{}); ok {
-				if len(confList) > 1 {
-					err = fmt.Errorf("Response contains multiple values: %v", len(confList))
-					log.Printf("[WARN] Response contains multiple values.")
-					return
-				}
-				if confMap, ok := confList[0].(map[string]interface{}); ok {
-					mapTmp = confMap
-					return
-				}
-			} else {
-				log.Printf("[WARN] Could not parse response type: %T", result[rspKey])
-			}
-		}
-	}
-	return
-}
-
-func delete(c *FortiSDKClient, method string, path string) (err error) {
-	result, string_body, err := sendRequest(c, method, path, nil)
-	if err != nil {
-		return err
-	}
-	err = fortiAPIErrorFormat(result, string_body)
-	return
-}
-
-func read(c *FortiSDKClient, method string, path string, bcomplex bool, rspKey string, params *map[string]interface{}) (mapTmp map[string]interface{}, err error) {
-	result, string_body, err := sendRequest(c, method, path, params)
 	if err != nil {
 		return nil, err
 	}
-	if fortiAPIHttpStatus404Checking(result) == true {
-		mapTmp = nil
-		return
+	mapTmp := map[string]interface{}{}
+	if rspKey == "entitlements" && result[rspKey] == nil { // TMP FortiFlex API BUG fix
+		rspKey = "vms"
 	}
-	err = fortiAPIErrorFormat(result, string_body)
-	if err == nil { //!!! need check whether all response is in rspKey argument
-		mapTmp = map[string]interface{}{}
-		if bcomplex {
-			new_rspKey := rspKey
-			if rspKey == "entitlements" && result[rspKey] == nil { // TMP FortiFlex API BUG fix
-				new_rspKey = "vms"
+	if result[rspKey] != nil {
+		if confMap, ok := result[rspKey].(map[string]interface{}); ok {
+			return confMap, nil
+		} else if confList, ok := result[rspKey].([]interface{}); ok {
+			if len(confList) > 1 {
+				err = fmt.Errorf("Response contains multiple values: %v", len(confList))
+				log.Printf("[WARN] Response contains multiple values.")
+				return nil, err
 			}
-			mapTmp[rspKey] = result[new_rspKey]
+			if confMap, ok := confList[0].(map[string]interface{}); ok {
+				return confMap, nil
+			}
 		} else {
-			mapTmp = (result[rspKey].([]interface{}))[0].(map[string]interface{})
+			log.Printf("[WARN] Could not parse response type: %T", result[rspKey])
 		}
 	}
-	return
+	return mapTmp, nil
 }
 
-func sendRequest(c *FortiSDKClient, method string, path string, params *map[string]interface{}) (result map[string]interface{}, string_body string, err error) {
+func read(client *FortiSDKClient, method string, path string, rspKey string, params *map[string]interface{}) (map[string]interface{}, error) {
+	result, string_body, err := sendRequest(client, method, path, params)
+	if err != nil {
+		return nil, err
+	}
+	err = fortiAPIErrorFormat(result, string_body)
+	if err != nil {
+		return nil, err
+	}
+	mapTmp := map[string]interface{}{}
+	new_rspKey := rspKey
+	if rspKey == "entitlements" && result[rspKey] == nil { // TMP FortiFlex API BUG fix
+		new_rspKey = "vms"
+	}
+	mapTmp[rspKey] = result[new_rspKey]
+	return mapTmp, nil
+}
+
+func sendRequest(client *FortiSDKClient, method string, path string, params *map[string]interface{}) (result map[string]interface{}, string_body string, err error) {
 	var locJSON []byte
 	if params != nil {
 		locJSON, err = json.Marshal(params)
-		log.Printf("FortiFlex request %s: %s", path, string(locJSON))
 		if err != nil {
 			log.Fatal(err)
 			return nil, "", err
@@ -93,8 +77,9 @@ func sendRequest(c *FortiSDKClient, method string, path string, params *map[stri
 		if locJSON != nil {
 			bytePara = bytes.NewBuffer(locJSON)
 		}
-		req := c.NewRequest(method, path, nil, bytePara)
-		err = req.Send()
+		log.Printf("[INFO] Request '%s' | %s", path, string(locJSON))
+		req := request.NewRequest(client.Auth, client.HTTPCon, method, path, nil, bytePara)
+		err = req.Send(5) // If the connection fails, retry up to 5 times
 		if err != nil || req.HTTPResponse == nil {
 			err = fmt.Errorf("cannot send request: %v", err)
 			return nil, "", err
@@ -108,10 +93,10 @@ func sendRequest(c *FortiSDKClient, method string, path string, params *map[stri
 		}
 
 		json.Unmarshal([]byte(string(body)), &result)
-		if result["status"] != nil {
+		if result["status"] != nil { // Retry for FortiFlex API error
 			rtStatus := fmt.Sprintf("%v", result["status"])
 			if rtStatus != "0" {
-				log.Printf("[Error] error response, will resend again %s, %v", path, result)
+				log.Printf("[ERROR] Response '%s' | retry time %v | %v", path, retry, result)
 				retry++
 				time.Sleep(time.Second)
 				continue
@@ -119,6 +104,6 @@ func sendRequest(c *FortiSDKClient, method string, path string, params *map[stri
 		}
 		break
 	}
-	log.Printf("FortiFlex reading %s: %s", path, string(body))
+	log.Printf("[INFO] Response '%s' | %s", path, string(body))
 	return result, string(body), err
 }