+++ test/lint_check.sh
@@ -1,6 +1,0 @@
-p=$(dirname "$PWD");
-export GOPATH=${p%/*/*/*/*}"/"
-make -C ../  fmt
-golint ../flexvm
-make -C ../  build
-